Move `Device` to `device_channel`
This allows us to enforce visibilty rules on its fields.
